+++ src/us/kbase/workspace/kbase/InitConstants.java
@@ -2,7 +2,6 @@ package us.kbase.workspace.kbase;
 
 public class InitConstants {
 	
-	public static final int TOKEN_REFRESH_INTERVAL_SEC = 24 * 60 * 60;
 	public static final String COL_SETTINGS = "settings";
 	public static final String COL_SHOCK_NODES = "shock_nodeMap";
 }